==> 1_Split_train_validation.ipynb <==
import pandas as pd
from sklearn.model_selection import train_test_split

df_train = pd.read_csv("data/train.csv")

display(df_train)

df_train.shape

features = df_train.drop(columns = ['price'])
target = df_train['price']

X_train, X_val, y_train, y_val = train_test_split(features, target, test_size = 0.2, random_state = 16)

display(X_train)

X_train.shape

X_val.shape

X_train.to_csv("split_data/train_features.csv", index=False)
y_train.to_csv("split_data/train_target.csv", index=False)

X_val.to_csv("split_data/val_features.csv", index=False)
y_val.to_csv("split_data/val_target.csv", index=False)
