==> 2_Preprocessing.ipynb <==
import pandas as pd
import numpy as np
import pickle
from datetime import datetime
from sklearn.model_selection import train_test_split
from sklearn.ensemble import IsolationForest
from ydata_profiling import ProfileReport

pd.set_option('display.max_columns', None)

X_train = pd.read_csv("split_data/train_features.csv")
y_train = pd.read_csv("split_data/train_target.csv", names=["price"], header = 0)
y_train.head()

df_train = pd.concat([X_train, y_train], axis = 1)
df_train.shape

# df_train['row_prop_missing'] = df_train.isna().mean(axis=1) # Not ok to already calculate here - some are missing by design (e.g. sticker)

profile = ProfileReport(df_train)
profile.to_notebook_iframe()

Observations  / remarks:
- id: make sure not to include
- is_appartment: multicollinearity with subtype -> don't include both
- area: high correlation with price! impute missing values based on median per subtype, add column to indicate original was missing
- added_time: number of minutes / seconds since the property was added? Can be relevant (very expensive properties might take longer before being sold)
- bedrooms: high correlation with area -> price!
- new_building: no remarks
- postcode: no remarks
- lat / lon: impute missings? calculate distance to railway track, airport, highway (noise)?
- advertiser: impute missings or treat as separate category? > 2000 unique values: reduce? 
- foto_amount: no remarks
- is_promoted: constant -> ignore
- subtype: limited number of missings, can be partly imputed based on is_appartment
- sticker (new / price drop): only price drop could be relevant for price? recode missing values to third category 'no sticker'? ! perfect correlation with energy_value?
- price drop date: relevance?
- energy_value: many missings! can be partly imputed using new_building and/or energy_label? note: a missing energy value usually means it's a bad one
- energy_label: many missings! multicollinearity with energy_value - value is more precise 
- province: maybe cross-check with postal code (DQ); use statbel average price per house / appartment data?
- price: maximum 999999?
- Outliers: Treat or not? DT-based methods can handle them

## Rows with many NAs

No removal done so far 

df_train = df_train.drop(['sticker'], axis = 1)

df_train['row_prop_missing'] = df_train.isna().mean(axis=1)

display(df_train[df_train['row_prop_missing'] > .25])

## Duplicate rows

# Check done: row with most recent (max) added_time does not necessarily correspond to row with lowest price

df_deduplicate = df_train.groupby(["area", "bedrooms", "postcode", "lat", "lon", "advertiser", "subtype"], dropna = False).agg({'added_time': 'max', 'id': 'nunique'}).reset_index()
df_deduplicate = df_deduplicate.rename(columns = {"id" : "row_count"})
df_deduplicate.head()

# Deduplicate based on added_time: keep most recent entry

df_train = pd.merge(df_train, df_deduplicate, how = 'inner', on = ["area", "bedrooms", "postcode", "lat", "lon", "advertiser", "subtype", "added_time"])
df_train = df_train.sort_values(by='added_time', ascending=False)
df_train['row_num'] = df_train.groupby(["area", "bedrooms", "postcode", "lat", "lon", "advertiser", "subtype"], dropna = False).cumcount() + 1
df_train = df_train[df_train['row_num'] == 1].drop(columns='row_num')
display(df_train)

df_train['id'].nunique()

## Subtype

Observations / remarks:
- Missing values are partly houses and partly appartments -> can be assigned accordingly
- There are synonyms -> can be grouped together
- 'Andere' is never an appartment
- Group infrequent levels together? Reduce dimensionality (but: not really necessary for DT-based approach?)
- Perform clustering a.o. on price to reduce number of levels?

df_train['subtype'].nunique()

pd.crosstab(df_train['subtype'], df_train['is_appartment'], dropna=False)

# Regroup infrequent categories as 'other'

subtype_counts = df_train['subtype'].value_counts()
min_subtype_count = 10
map_infrequent_subtypes = subtype_counts[subtype_counts < min_subtype_count].index

df_train['subtype_regrouped'] = df_train['subtype'].apply(lambda x: 'Andere' if x in map_infrequent_subtypes else x) # remove this step?

# Save infrequent categories to recode them in validation and test set

with open('intermediate_data/map_infrequent_subtypes.pkl', 'wb') as file:
    pickle.dump(map_infrequent_subtypes, file)

# Recode subtypes

map_synonyms = {
    'Assistentie-appartement': 'Serviceflat',
    'Villa-landhuis': 'Villa',
    'Moderne villa': 'Villa',
    'Eengezinswoning': 'Woning',    
    # 'Herenwoning': 'Herenhuis',
    # 'Dakappartement': 'Penthouse',
    'Studio met slaaphoek': 'Studio',
    # 'Rijwoning': 'Woning',
    'Rijwoning': 'Rijwoning',
    'Gelijkvloers app.': 'Appartement',
    'Uitzonderlijke woning': 'Villa',
    'Herenwoning': 'Villa',
    'Herenhuis': 'Villa',
    'Burgerswoning': 'Woning',
    'Koppelwoning': 'Woning',
    'Duplex': 'Appartement',
    'Triplex': 'Appartement',
    'Bungalow': 'Woning',
    'Hoeve': 'Villa',
    'Fermette': 'Woning',
    'Bel-étage': 'Woning',
    'Hoekwoning': 'Woning',
    'Pastorijwoning': 'Woning',
    'Arbeiderswoning': 'Woning',
    'Loft': 'Loft Penthouse',
    'Dakappartement': 'Appartement',
    'Penthouse': 'Loft Penthouse',
    'Chalet': 'Andere',
    'Cottage': 'Andere',
    'Vakantiewoning': 'Andere',    
    'Gemengd gebruik': 'Appartementsgebouw' # More similar to this category than to "others"
    
}

df_train['subtype_regrouped'] = df_train['subtype_regrouped'].replace(map_synonyms)
df_train['subtype_regrouped'] = df_train['subtype_regrouped'].fillna('Andere')

df_train['subtype_regrouped'].value_counts()

# Save mapping to use in validation and test data

with open('intermediate_data/map_synonyms.pkl', 'wb') as file:
     pickle.dump(map_synonyms, file)

# Get median price per regrouped subtype to use as feature

subtype_median_price = df_train.groupby('subtype_regrouped')['price'].median().reset_index()
subtype_median_price.columns = ['subtype_regrouped', 'subtype_median_price']
subtype_median_price

# Save median price per regrouped subtype to use as feature in validation and test set

subtype_median_price.to_pickle('intermediate_data/subtype_median_price.pkl')

df_train = pd.merge(df_train, subtype_median_price, how = 'left', on = 'subtype_regrouped')
display(df_train)

## Area

Impute area with median value for subtype (regrouped)

df_train['area_missing'] = df_train['area'].isna().astype(int)
df_train['area_missing'].value_counts(dropna = False)

df_train.groupby('is_appartment')['area'].median()

df_train.groupby('is_appartment')['area'].max()

# Get median area per subtype and province combination
# Save to use for imputation in validation and test set

median_area = df_train.groupby(['subtype_regrouped', 'province'])['area'].median()
display(median_area)
median_area.to_pickle('intermediate_data/median_area.pkl')

# Aggregate only based on subtype because some combinations with province result in missings and differences between provinces are small

median_area = df_train.groupby('subtype_regrouped')['area'].median()
display(median_area)
median_area.to_pickle('intermediate_data/median_area.pkl')

# Attempt 3: also aggregate based on number of bedrooms

median_area_bedrooms = df_train.groupby(['subtype_regrouped', 'bedrooms'])['area'].median()
display(median_area_bedrooms)
median_area_bedrooms.to_pickle('intermediate_data/median_area_bedrooms.pkl')

df_train['area_imputed1'] = df_train.apply(
    lambda row: median_area_bedrooms.get((row['subtype_regrouped'], row['bedrooms']), row['area']) if pd.isna(row['area']) else row['area'],
    axis=1
)
df_train['area_imputed2'] = df_train.apply(
    lambda row: median_area.get((row['subtype_regrouped']), row['area']) if pd.isna(row['area']) else row['area'],
    axis=1
)

df_train['area_imputed'] = df_train['area_imputed1'].combine_first(df_train['area_imputed2'])

df_train.drop(['area_imputed1', 'area_imputed2'], axis = 1).head()

# No more missing values for area

display(df_train[df_train['area_imputed'].isna()])

# Calculate area relative to number of bedrooms (+1) to add as potential feature

df_train['area_rel_to_bedrooms'] = df_train['area_imputed'] / (df_train['bedrooms'] + 1)

## Energy value

df_train['energy_label'].value_counts(dropna = False)

New buildings with energy label d, e, f, g -> probably bad DQ?

pd.crosstab(df_train['energy_label'], df_train['new_building'], dropna=False)

# Keep only first letter of energy label, except for 'a+'

df_train['energy_label_regrouped'] = df_train['energy_label'].apply(lambda x: x[0] if isinstance(x, str) and x != 'a+' and x != 'a+' else x)

pd.crosstab(df_train['energy_label_regrouped'], df_train['new_building'], dropna=False)

For f and g not ok: normally label A corresponds to values 0-100, B to 101-200, etc.

df_train.groupby(['energy_label_regrouped'])['energy_value'].median()

df_train.groupby(['energy_label_regrouped', 'new_building'])['energy_value'].median()

df_train[df_train['energy_value'].isna()]

# Count NAs for energy label x new building

pd.crosstab(df_train[df_train['energy_value'].isna()]['energy_label_regrouped'], df_train[df_train['energy_value'].isna()]['new_building'], dropna=False)

For categories with missings, the sample size with valid energy values is large enough (see higher), so it's okay to also impute based on new_building

# Get median energy value per energy label x new building x subtype combination
# Save to use for imputation in validation and test set

median_energy = df_train.groupby(['energy_label_regrouped', 'new_building', 'subtype_regrouped'])['energy_value'].median()

median_energy.to_pickle('intermediate_data/median_energy.pkl')

df_train['energy_value_missing'] = df_train['energy_value'].isna().astype(int)

df_train['energy_value_imputed'] = df_train.apply(
    lambda row: median_energy.get((row['energy_label_regrouped'], row['new_building'], row['subtype_regrouped']), row['energy_value']) if pd.isna(row['energy_value']) else row['energy_value'],
    axis=1
)

# Check how many cases still have missing energy value (cases without energy label)

df_train['energy_value'].value_counts(dropna = False)

# Now all rows except for the ones where the energy label is also missing have an energy value

df_train['energy_value_imputed'].value_counts(dropna = False)

df_train.groupby(['subtype_regrouped', 'new_building'])['energy_value'].median()

median_energy_wo_label = df_train.groupby(['new_building', 'subtype_regrouped'])['energy_value'].median()

median_energy_wo_label.to_pickle('intermediate_data/median_energy_wo_label.pkl')

df_train['energy_value_imputed'] = df_train.apply(
    lambda row: median_energy_wo_label.get((row['new_building'], row['subtype_regrouped']), row['energy_value']) if pd.isna(row['energy_value_imputed']) else row['energy_value_imputed'],
    axis=1
)

# All NAs are imputed

df_train['energy_value_imputed'].value_counts(dropna = False)

## Advertiser

High number of categories - apply something similar to Weights Of Evidence (= for classification problems)

Note: some advertisers occur only once and seem to have a person's name (not a real estate agency) -> also informative

Make bins of advertisers based on how often they occur (only once (category 5: person) vs. more (categories 1-4: agency)) and their median price.
Category 5: only occurs once, so no relevant information on 'typical' pricing; this advertiser will (normally) not occur in the validation or test set either 
Catgories 1-4: occurs more than once, categorize based on median price 

df_train['advertiser'].nunique()

df_train['advertiser'].value_counts(dropna = False).head(50) 

median_price_advertiser = df_train.groupby('advertiser', dropna = False).agg(
    advertiser_median_price=('price', 'median'),  
    advertiser_count=('advertiser', 'size')     
).reset_index()

median_price_advertiser.columns = ['advertiser', 'advertiser_median_price', 'advertiser_count']
display(median_price_advertiser)

median_price_advertiser['advertiser_bin'] = np.where((median_price_advertiser['advertiser_count'] == 1) | (pd.isna(median_price_advertiser['advertiser']))
                                                     , 5, np.nan)
display(median_price_advertiser)

num_bins = 4  
median_price_advertiser.loc[(median_price_advertiser['advertiser_count'] > 1) & (~pd.isna(median_price_advertiser['advertiser'])), 'advertiser_bin'] = pd.qcut(
    median_price_advertiser.loc[(median_price_advertiser['advertiser_count'] > 1) & ~ (pd.isna(median_price_advertiser['advertiser'])), 'advertiser_median_price'], 
    q=num_bins, 
    labels=False
)
display(median_price_advertiser)

median_price_advertiser.groupby(['advertiser_bin'])['advertiser_median_price'].median()

with open('intermediate_data/median_price_advertiser.pkl', 'wb') as file:
    pickle.dump(median_price_advertiser, file)

df_train = pd.merge(df_train, median_price_advertiser, how='left', on = "advertiser")
df_train.head()

df_train['advertiser_bin'].value_counts(dropna = False)

median_price_per_advertiser_bin = df_train.groupby(['advertiser_bin'], as_index = False)['price'].median()
median_price_per_advertiser_bin.columns = ['advertiser_bin', 'median_price_advertiser_bin']
median_price_per_advertiser_bin

median_price_per_advertiser_bin.to_pickle('intermediate_data/median_price_per_advertiser_bin.pkl')

df_train = pd.merge(df_train, median_price_per_advertiser_bin, how='left', on='advertiser_bin')
df_train.head()

## Regional prices - Statbel

# File to map NIS codes to zipcodes

df_map_nis = pd.read_csv("external_data/cities.csv")
df_map_nis.head()

df_map_nis['zipCode'].value_counts()

df_map_nis[df_map_nis['zipCode'] == 3020] # if there are multiple lines for the same zip code, there is 1 'main'

df_map_nis_renamed = df_map_nis.rename(columns={'name': 'municipality', 'province': 'province_nis', 'zipCode': 'zip_code', 'nisCode': 'nis_code'}).copy()

# Deduplicate NIS codes based on 'main' column

df_map_nis_deduplicated = df_map_nis_renamed.sort_values(by='main', ascending=False).groupby('zip_code').first().reset_index().drop(columns = 'main').copy()

df_map_nis_deduplicated['zip_code'].value_counts()

with open('intermediate_data/map_nis_deduplicated.pkl', 'wb') as file:
    pickle.dump(df_map_nis_deduplicated, file)

# Add the NIS data to the training data

df_train = pd.merge(df_train, df_map_nis_deduplicated, how='left', left_on='postcode', right_on = 'zip_code')
display(df_train)

df_train.shape

# Check merge: do provinces in training data correspond to provinces in NIS data

pd.crosstab(df_train['province'], df_train['province_nis'], dropna=False) # looks okay except for Antwerpen-Oost-Vlaanderen?

# Get Statbel real estate sales prices per NIS code for 2024

df_statbel_prices = pd.read_excel("external_data/vastgoed_2010_9999.xlsx", sheet_name = '2024')
df_statbel_prices.head()

df_statbel_prices['CD_TYPE_NL'].unique()

# Df with Statbel data for houses for 1st semester of 2024

df_statbel_prices_houses = (
    df_statbel_prices
    .query("CD_TYPE_NL == 'Alle huizen met 2, 3, 4 of meer gevels (excl. appartementen)' & CD_PERIOD == 'S1'")  
    .loc[:, ['CD_REFNIS', 'CD_REFNIS_NL', 'MS_TOTAL_TRANSACTIONS', 'MS_P_25', 'MS_P_50_median', 'MS_P_75']]  
    .assign(F_APPARTMENT = 0)
)

# Df with Statbel data for appartments for 1st semester of 2024

df_statbel_prices_appartments = (
    df_statbel_prices
    .query("CD_TYPE_NL == 'Appartementen' & CD_PERIOD == 'S1'")  
    .loc[:, ['CD_REFNIS', 'CD_REFNIS_NL', 'MS_TOTAL_TRANSACTIONS', 'MS_P_25', 'MS_P_50_median', 'MS_P_75']]  
    .assign(F_APPARTMENT = 1)
)

# Stack both dfs

df_statbel_prices_selection = pd.concat([df_statbel_prices_houses, df_statbel_prices_appartments], ignore_index=True)

display(df_statbel_prices_selection.head())

df_statbel_prices_selection[df_statbel_prices_selection['CD_REFNIS_NL'].str.contains('PROVINCIE')]

with open('intermediate_data/statbel_prices_selection.pkl', 'wb') as file:
    pickle.dump(df_statbel_prices_selection, file)

# Example of NIS where there are no appartment prices available

df_statbel_prices[df_statbel_prices['CD_REFNIS'] == 13049].head()

# Add Statbel data to training data 

df_train = pd.merge(df_train, df_statbel_prices_selection, how='left', left_on=['nis_code', 'is_appartment'], right_on = ['CD_REFNIS', 'F_APPARTMENT'])
display(df_train)

# Quite some missings, whereas profiling below suggests it's a highly relevant feature

df_train['MS_P_50_median'].isna().sum()

# Add zip codes to prices data in order to aggregate and use for imputation

df_statbel_prices_selection = pd.merge(df_statbel_prices_selection, df_map_nis_deduplicated, how='left', left_on=['CD_REFNIS'], right_on = ['nis_code'])
display(df_statbel_prices_selection)

# Calculate median of MS_P_50_median per first 2 digits of zipcode and house vs. appartement for imputation

df_statbel_prices_selection["zip_code_first2"] = df_statbel_prices_selection.zip_code.astype(str).str[:2]
median_price_per_zipcode = df_statbel_prices_selection.groupby(['zip_code_first2', 'F_APPARTMENT'], as_index = False)['MS_P_50_median'].median()
median_price_per_zipcode.columns = ['zip_code_first2', 'is_appartment', 'med_MS_P_50_median']
display(median_price_per_zipcode.head(20))

median_price_per_zipcode['med_MS_P_50_median'].isna().sum()

# Make dict of the table above to save and use for imputation on validation and test set

price_lookup_zipcode = median_price_per_zipcode.set_index(
    ['zip_code_first2', 'is_appartment']
)['med_MS_P_50_median'].to_dict()

with open('intermediate_data/price_lookup_zipcode.pkl', 'wb') as file:
    pickle.dump(price_lookup_zipcode, file)

df_train["zip_code_first2"] = df_train.postcode.astype(str).str[:2]

df_train["MS_P_50_median_imputed"] = df_train.apply(
    lambda row: price_lookup_zipcode.get(
        (row["zip_code_first2"], row["is_appartment"]),
        row["MS_P_50_median"]
    ) if pd.isna(row["MS_P_50_median"]) else row["MS_P_50_median"],
    axis=1
)

# Still quite some missings

df_train['MS_P_50_median_imputed'].isna().sum()

df_train[df_train['MS_P_50_median_imputed'].isna()].head()

# Calculate median of MS_P_50_median per first digit of zipcode and house vs. appartement for imputation

df_statbel_prices_selection["zip_code_first1"] = df_statbel_prices_selection.zip_code.astype(str).str[:1]
median_price_per_province = df_statbel_prices_selection.groupby(['zip_code_first1', 'F_APPARTMENT'], as_index = False)['MS_P_50_median'].median()
median_price_per_province.columns = ['zip_code_first1', 'is_appartment', 'med_MS_P_50_median']
display(median_price_per_province.head(20))

price_lookup_province = median_price_per_province.set_index(
    ['zip_code_first1', 'is_appartment']
)['med_MS_P_50_median'].to_dict()

with open('intermediate_data/price_lookup_province.pkl', 'wb') as file:
    pickle.dump(price_lookup_province, file)

df_train["zip_code_first1"] = df_train.postcode.astype(str).str[:1]

df_train["MS_P_50_median_imputed"] = df_train.apply(
    lambda row: price_lookup_province.get(
        (row["zip_code_first1"], row["is_appartment"]),
        row["MS_P_50_median"]
    ) if pd.isna(row["MS_P_50_median_imputed"]) else row["MS_P_50_median_imputed"],
    axis=1
)

# All missings have been imputed

df_train[df_train['MS_P_50_median_imputed'].isna()]

## Regional prices - Price per area 

# Calculate price per squared meter and aggregate per first 2 zipcode digits

df_train['price_per_area'] = df_train['price'] / df_train['area']
price_per_area_per_region = df_train.groupby('zip_code_first2')['price_per_area'].median().round(2).reset_index()
price_per_area_per_region.columns = ['zip_code_first2', 'median_price_per_area']
display(price_per_area_per_region.head(30))

# Save for use in validation and test data

with open('intermediate_data/price_per_area_per_region.pkl', 'wb') as file:
    pickle.dump(price_per_area_per_region, file)

df_train = pd.merge(df_train, price_per_area_per_region, how = 'left', on = 'zip_code_first2')

# Do the same as above but with the first 3 zipcode digits (= more granular)

df_train["zip_code_first3"] = df_train.postcode.astype(str).str[:3]

df_train['price_per_area_2'] = df_train['price'] / df_train['area']
price_per_area_per_region_2 = df_train.groupby('zip_code_first3')['price_per_area'].median().round(2).reset_index()
price_per_area_per_region_2.columns = ['zip_code_first3', 'median_price_per_area_2']
display(price_per_area_per_region_2.head(30))

# Save for use in validation and test data

with open('intermediate_data/price_per_area_per_region_2.pkl', 'wb') as file:
    pickle.dump(price_per_area_per_region_2, file)

# Add to training data, prioritizing the amount on the 3 digit level

df_train = pd.merge(df_train, price_per_area_per_region_2, how = 'left', on = 'zip_code_first3')

df_train['median_price_per_area_2'] = df_train['median_price_per_area_2'].combine_first(df_train['median_price_per_area'])

## Regional prices - replace Statbel

Use training data rather than external resource that needs to be updated over time

# Aggregate median prices per zipcode

price_per_area_type_zipcode = df_train.groupby(['postcode', 'is_appartment'])['price'].median().round(2).reset_index()

price_per_area_type_zipcode.columns = ['postcode', 'is_appartment', 'price_per_area_type_zipcode']
display(price_per_area_type_zipcode)

# Aggregate median prices per first 3 digits of zipcode (in case of missings on zipcode level)

df_train["postcode_first3"] = df_train.postcode.astype(str).str[:3]

price_per_area_type_zipcodefirst3 = df_train.groupby(['postcode_first3', 'is_appartment'])['price'].median().round(2).reset_index()
price_per_area_type_zipcodefirst3.columns = ['postcode_first3', 'is_appartment', 'price_per_area_type_zipcodefirst3']
display(price_per_area_type_zipcodefirst3)

# Aggregate median prices per first 2 digits of zipcode (in case of missings on first 3 digits level)

df_train["postcode_first2"] = df_train.postcode.astype(str).str[:2]

price_per_area_type_zipcodefirst2 = df_train.groupby(['postcode_first2', 'is_appartment'])['price'].median().round(2).reset_index()
price_per_area_type_zipcodefirst2.columns = ['postcode_first2', 'is_appartment', 'price_per_area_type_zipcodefirst2']
display(price_per_area_type_zipcodefirst2)

# Aggregate median prices per first digit of zipcode (in case of missings on first 2 digits level)

df_train["postcode_first1"] = df_train.postcode.astype(str).str[:1]

price_per_area_type_zipcodefirst1 = df_train.groupby(['postcode_first1', 'is_appartment'])['price'].median().round(2).reset_index()
price_per_area_type_zipcodefirst1.columns = ['postcode_first1', 'is_appartment', 'price_per_area_type_zipcodefirst1']
display(price_per_area_type_zipcodefirst1)

# Save all aggregations above for use in validation and test set

with open('intermediate_data/price_per_area_type_zipcode.pkl', 'wb') as file:
    pickle.dump(price_per_area_type_zipcode, file)

with open('intermediate_data/price_per_area_type_zipcodefirst3.pkl', 'wb') as file:
    pickle.dump(price_per_area_type_zipcodefirst3, file)

with open('intermediate_data/price_per_area_type_zipcodefirst2.pkl', 'wb') as file:
    pickle.dump(price_per_area_type_zipcodefirst2, file)

with open('intermediate_data/price_per_area_type_zipcodefirst1.pkl', 'wb') as file:
    pickle.dump(price_per_area_type_zipcodefirst1, file)

# Add median prices per zip code to training data, prioritizing most granular zipcode level

df_train = pd.merge(df_train, price_per_area_type_zipcode, how = 'left', on = ['postcode', 'is_appartment'])
df_train = pd.merge(df_train, price_per_area_type_zipcodefirst3, how = 'left', on = ['postcode_first3', 'is_appartment'])
df_train = pd.merge(df_train, price_per_area_type_zipcodefirst2, how = 'left', on = ['postcode_first2', 'is_appartment'])
df_train = pd.merge(df_train, price_per_area_type_zipcodefirst1, how = 'left', on = ['postcode_first1', 'is_appartment'])
df_train['price_per_area_type_region'] = df_train['price_per_area_type_zipcode'].\
combine_first(df_train['price_per_area_type_zipcodefirst3']).\
combine_first(df_train['price_per_area_type_zipcodefirst2']).\
combine_first(df_train['price_per_area_type_zipcodefirst1'])
display(df_train[df_train['price_per_area_type_zipcode'].isna()])

display(df_train[df_train['price_per_area_type_region'].isna()])

## Lat / Lon

df_train[df_train['lon'].isna()]

# Add flag indicating lat / lon is missing

df_train['lat_missing'] = df_train['lat'].isna().astype(int)
df_train['lon_missing'] = df_train['lon'].isna().astype(int)

Attempt 3: remove outlying coordinates
<br> belgium.be: Belgium spans 2 degrees in latitude, from 51 degrees 30 minutes N at Meerle (northernmost point) to 49 degrees 30 minutes N at Torgny (southernmost point). In longitude, it spans less than 4 degrees, from 2 degrees 33 minutes E to 6 degrees 24 minutes E.

df_train[(df_train['lat'] < 49.5) | (df_train['lat'] > 51.5) | (df_train['lon'] < 2.55) | (df_train['lon'] > 6.4)]

df_train = df_train[(df_train['lat'].isna()) | (df_train['lon'].isna())  | ((df_train['lat'] >= 49.5) & (df_train['lat'] <= 51.5) & (df_train['lon'] >= 2.55) & (df_train['lon'] <= 6.4))]

# Calculate mean latitude for imputation and save for use on validation and test data

mean_lat = df_train.groupby('province')['lat'].mean()
mean_lat.to_pickle('intermediate_data/mean_lat.pkl')
display(mean_lat)

# Calculate mean longitude for imputation and save for use on validation and test data

mean_lon = df_train.groupby('province')['lon'].mean()
mean_lon.to_pickle('intermediate_data/mean_lon.pkl')

# Impute lat and lon

df_train['lat_imputed'] = df_train.apply(
    lambda row: mean_lat.get((row['province']), row['lat']) if pd.isna(row['lat']) else row['lat'],
    axis=1
)

df_train['lon_imputed'] = df_train.apply(
    lambda row: mean_lon.get((row['province']), row['lon']) if pd.isna(row['lon']) else row['lon'],
    axis=1
)

display(df_train)

df_train[df_train['lat_imputed'].isna()]

## Price drop flag

Add a flag indicating if the price for a certain property has been dropped

df_train['price_dropped'] = abs(df_train['price_drop_date'].isna().astype(int) - 1)
display(df_train)

## Zipcode last digits

If last 3 digits are 0: big city
If last 2 but not last 3 digits are 0: city

df_train["postcode_last3"] = df_train.postcode.astype(str).str[1:] 
df_train["postcode_last2"] = df_train.postcode.astype(str).str[2:]
df_train["postcode_last3_0"] = (df_train["postcode_last3"] == '000').astype('int')
df_train["postcode_last2_0"] = ((df_train["postcode_last3"] != '000') & (df_train["postcode_last2"] == '00')).astype('int')
display(df_train)

## Profile again

# Leave out percentiles 25 and 75 for prices per NIS code because of multicollinearity

df_train_sel = df_train[['bedrooms', 'new_building', 'foto_amount', 'province', 'subtype_regrouped', 'subtype_median_price', 'area_missing', 'area_imputed', 'area_rel_to_bedrooms', 'energy_value_missing', 'energy_value_imputed', 'advertiser_count', 'median_price_advertiser_bin', 'MS_TOTAL_TRANSACTIONS', 'MS_P_50_median_imputed', 'median_price_per_area', 'price']].drop_duplicates()

profile_sel = ProfileReport(df_train_sel)

profile_sel.to_notebook_iframe()

## Final variable selection and encoding


df_train_sel = df_train[['new_building', 'foto_amount', 'province', 'subtype_median_price', 'area_missing', 'area_imputed', 'energy_value_missing', 'energy_value_imputed', 'advertiser_count', 'median_price_advertiser_bin', 'MS_P_50_median_imputed', 'median_price_per_area', 'price']].drop_duplicates()

display(df_train_sel)

X_train_preprocessed = df_train_sel.drop('price', axis = 1)
X_train_preprocessed = pd.get_dummies(X_train_preprocessed, columns=['province'], drop_first=True)

display(X_train_preprocessed)

y_train_preprocessed = df_train_sel['price']

X_train_preprocessed.to_csv("split_data/train_features_preprocessed.csv", index=False)
y_train_preprocessed.to_csv("split_data/train_target_preprocessed.csv", index=False)

X_train_preprocessed.columns

## Final variable selection and encoding - lat/lon instead of province

df_train_sel_2 = df_train[['new_building', 'foto_amount', 'lat_missing', 'lat_imputed', 'lon_missing', 'lon_imputed', 'subtype_median_price', 'area_missing', 'area_imputed', 'energy_value_missing', 'energy_value_imputed', 'advertiser_count', 'median_price_advertiser_bin', 'MS_P_50_median_imputed', 'median_price_per_area', 'price']].drop_duplicates()

profile_sel_2 = ProfileReport(df_train_sel_2)

profile_sel_2.to_notebook_iframe()

X_train_preprocessed_2 = df_train_sel_2.drop('price', axis = 1)

y_train_preprocessed_2 = df_train_sel_2['price']

X_train_preprocessed_2.to_csv("split_data/train_features_preprocessed_2.csv", index=False)
y_train_preprocessed_2.to_csv("split_data/train_target_preprocessed_2.csv", index=False)

## Final variable selection and encoding - attempt 3a

df_train_sel_3 = df_train[['new_building', 'foto_amount', 'lat_missing', 'lat_imputed', 'lon_imputed', 'subtype_median_price', 'area_missing', 'area_imputed', 'energy_value_missing', 'energy_value_imputed', 'advertiser_count', 'median_price_advertiser_bin', 'price_per_area_type_region', 'median_price_per_area', 'price_dropped', 'price']].drop_duplicates()

profile_sel_3 = ProfileReport(df_train_sel_3)

# Strong negative correlation between median price per subtype and median price per area, per type (is appartment Y/N) per region (zipcode)?
# No longer the case when switching back to median price (not median price per area) per type per region

profile_sel_3.to_notebook_iframe()

### Outlier detection

Detected but currently nothing removed 

model = IsolationForest(contamination=0.005)  # expected % of outliers
model.fit(df_train_sel_3[['area_imputed', 'energy_value_imputed', 'price']])
outliers = model.predict(df_train_sel_3[['area_imputed', 'energy_value_imputed', 'price']])  # returns 1 for inliers, -1 for outliers

display(df_train_sel_3[outliers == -1])

X_train_preprocessed_3 = df_train_sel_3.drop('price', axis = 1)
y_train_preprocessed_3 = df_train_sel_3['price']

X_train_preprocessed_3.to_csv("split_data/train_features_preprocessed_3.csv", index=False)
y_train_preprocessed_3.to_csv("split_data/train_target_preprocessed_3.csv", index=False)

## Final variable selection and encoding - attempt 3b

df_train_sel_3b = df_train[['new_building', 'foto_amount', 'lat_missing', 'lat_imputed', 'lon_imputed', 'subtype_median_price', 'area_missing', 'area_imputed', 'energy_value_missing', 'energy_value_imputed', 'advertiser_count', 'median_price_advertiser_bin', 'MS_P_50_median_imputed', 'median_price_per_area_2', 'postcode_last2_0', 'postcode_last3_0', 'price_dropped', 'price']].drop_duplicates()

profile_sel_3b = ProfileReport(df_train_sel_3b)

profile_sel_3b.to_notebook_iframe()

X_train_preprocessed_3b = df_train_sel_3b.drop('price', axis = 1)
y_train_preprocessed_3b = df_train_sel_3b['price']

X_train_preprocessed_3b.to_csv("split_data/train_features_preprocessed_3b.csv", index=False)
y_train_preprocessed_3b.to_csv("split_data/train_target_preprocessed_3b.csv", index=False)

## Smaller training set for TabPFN

df_train_sel_3 = df_train[['new_building', 'foto_amount', 'lat_missing', 'lat_imputed', 'lon_missing', 'lon_imputed', 'subtype_median_price', 'area_missing', 'area_imputed', 'energy_value_missing', 'energy_value_imputed', 'advertiser_count', 'median_price_advertiser_bin', 'MS_P_50_median_imputed', 'median_price_per_area', 'price']].drop_duplicates()

display(df_train_sel_3)

X_train_3 = df_train_sel_3.drop('price', axis = 1)

y_train_3 = df_train_sel_3['price']

X_train_3_1, X_train_3_2, y_train_3_1, y_train_3_2 = train_test_split(X_train_3, y_train_3, test_size=.67, random_state=16)

X_train_3_1.shape

y_train_3_1.shape

X_train_3_2.shape

X_train_3_2, X_train_3_3, y_train_3_2, y_train_3_3 = train_test_split(X_train_3_2, y_train_3_2, test_size=.5, random_state=16)

X_train_3_2.shape

y_train_3_2.shape

X_train_3_3.shape

y_train_3_3.shape

X_train_3_1.to_csv("split_data/train_features_preprocessed_3_1.csv", index=False)
y_train_3_1.to_csv("split_data/train_target_preprocessed_3_1.csv", index=False)
X_train_3_2.to_csv("split_data/train_features_preprocessed_3_2.csv", index=False)
y_train_3_2.to_csv("split_data/train_target_preprocessed_3_2.csv", index=False)
X_train_3_3.to_csv("split_data/train_features_preprocessed_3_3.csv", index=False)
y_train_3_3.to_csv("split_data/train_target_preprocessed_3_3.csv", index=False)

X_train_4_1, X_train_4_2, y_train_4_1, y_train_4_2 = train_test_split(X_train_3, y_train_3, test_size=.52, random_state=16)

X_train_4_1.shape

y_train_4_1.shape

X_train_4_2.shape

X_train_4_2, X_train_4_3, y_train_4_2, y_train_4_3 = train_test_split(X_train_4_2, y_train_4_2, test_size=.08, random_state=16)

X_train_4_2.shape

y_train_4_2.shape

X_train_4_1.to_csv("split_data/train_features_preprocessed_4_1.csv", index=False)
y_train_4_1.to_csv("split_data/train_target_preprocessed_4_1.csv", index=False)
X_train_4_2.to_csv("split_data/train_features_preprocessed_4_2.csv", index=False)
y_train_4_2.to_csv("split_data/train_target_preprocessed_4_2.csv", index=False)
X_train_4_3.to_csv("split_data/train_features_preprocessed_4_3.csv", index=False)
y_train_4_3.to_csv("split_data/train_target_preprocessed_4_3.csv", index=False)
