==> 5_Preprocessing_test_data.ipynb <==
import pandas as pd
import numpy as np
import pickle
from ydata_profiling import ProfileReport

pd.set_option('display.max_columns', None) # show all columns in a df

X_test = pd.read_csv("data/test.csv")

df_test = X_test

df_test.shape

df_test.head()

profile = ProfileReport(df_test)

profile.to_notebook_iframe()

## Rows with many NAs

# display(df_test[df_test['row_prop_missing'] > .10])

## Duplicate rows

## Subtype

Observations / remarks:
- Missing values are partly houses and partly appartments -> can be assigned accordingly
- There are synonyms -> can be grouped together
- 'Andere' is never an appartment
- Group infrequent levels together? Reduce dimensionality (but: not really necessary for DT-based approach?)
- Perform clustering a.o. on price to reduce number of levels?

pd.crosstab(df_test['subtype'], df_test['is_appartment'], dropna=False)

with open('intermediate_data/map_infrequent_subtypes.pkl', 'rb') as file:
    map_infrequent_subtypes = pickle.load(file)

df_test['subtype_regrouped'] = df_test['subtype'].apply(lambda x: 'Andere' if x in map_infrequent_subtypes else x) # remove this step?

map_synonyms = {
    'Assistentie-appartement': 'Serviceflat',
    'Villa-landhuis': 'Villa',
    'Moderne villa': 'Villa',
    'Eengezinswoning': 'Woning',    
    # 'Herenwoning': 'Herenhuis',
    # 'Dakappartement': 'Penthouse',
    'Studio met slaaphoek': 'Studio',

    # 'Rijwoning': 'Woning',
    'Gelijkvloers app.': 'Appartement',
    'Uitzonderlijke woning': 'Villa',
    'Herenwoning': 'Villa',
    'Herenhuis': 'Villa',
    'Burgerswoning': 'Woning',
    'Koppelwoning': 'Woning',
    'Duplex': 'Appartement',
    'Triplex': 'Appartement',
    'Bungalow': 'Woning',
    'Hoeve': 'Villa',
    'Fermette': 'Woning',
    'Bel-étage': 'Woning',
    'Hoekwoning': 'Woning',
    'Pastorijwoning': 'Woning',
    'Arbeiderswoning': 'Woning',
    'Loft': 'Loft Penthouse',
    'Dakappartement': 'Appartement',
    'Penthouse': 'Loft Penthouse',
    'Chalet': 'Andere',
    'Cottage': 'Andere',
    'Vakantiewoning': 'Andere',
    'Gemengd gebruik': 'Andere',
    'Woonboot': 'Andere', # To do: implement solution for new building types not seen in training data -> should become Other
    'Sociale woning': 'Andere' # To do: implement solution for new building types not seen in training data -> should become Other
    
    
}

df_test['subtype_regrouped'] = df_test['subtype_regrouped'].replace(map_synonyms)
df_test['subtype_regrouped'] = df_test['subtype_regrouped'].fillna('Andere')

df_test['subtype_regrouped'].value_counts()

with open('intermediate_data/subtype_median_price.pkl', 'rb') as file:
    subtype_median_price = pickle.load(file)

df_test = pd.merge(df_test, subtype_median_price, how = 'left', on = 'subtype_regrouped')
display(df_test)

## Area

Impute area with median value for subtype (regrouped) and province

df_test['area_missing'] = df_test['area'].isna().astype(int)

with open('intermediate_data/median_area.pkl', 'rb') as file:
    median_area = pickle.load(file)

df_test['area_imputed'] = df_test.apply(
    lambda row: median_area.get((row['subtype_regrouped']), row['area']) if pd.isna(row['area']) else row['area'],
    axis=1
)

df_test['area_rel_to_bedrooms'] = df_test['area_imputed'] / (df_test['bedrooms'] + 1)

## Energy value

# Keep only first letter of energy label, except for 'a+'

df_test['energy_label_regrouped'] = df_test['energy_label'].apply(lambda x: x[0] if isinstance(x, str) and x != 'a+' and x != 'a+' else x)

pd.crosstab(df_test['energy_label_regrouped'], df_test['new_building'], dropna=False)

For f and g not ok: normally label A corresponds to values 0-100, B to 101-200, etc.

df_test['energy_value_missing'] = df_test['energy_value'].isna().astype(int)

with open('intermediate_data/median_energy.pkl', 'rb') as file:
    median_energy = pickle.load(file)

df_test['energy_value_imputed'] = df_test.apply(
    lambda row: median_energy.get((row['energy_label_regrouped'], row['new_building'], row['subtype_regrouped']), row['energy_value']) if pd.isna(row['energy_value']) else row['energy_value'],
    axis=1
)

# Now all rows except for the ones where the energy label is also missing have an energy value

df_test['energy_value_imputed'].value_counts(dropna = False)

df_test.groupby(['subtype_regrouped', 'new_building'])['energy_value'].median()

with open('intermediate_data/median_energy_wo_label.pkl', 'rb') as file:
    median_energy_wo_label = pickle.load(file)

df_test['energy_value_imputed'] = df_test.apply(
    lambda row: median_energy_wo_label.get((row['new_building'], row['subtype_regrouped']), row['energy_value']) if pd.isna(row['energy_value_imputed']) else row['energy_value_imputed'],
    axis=1
)

# All NAs imputed

df_test['energy_value_imputed'].value_counts(dropna = False)

## Advertiser

High number of categories - apply something similar to Weights Of Evidence (= for classification problems)

Note: some advertisers occur only once and seem to have a person's name (not a real estate agency) -> also informative

Make bins of advertisers based on how often they occur (only once (category 5: person) vs. more (categories 1-4: agency)) and their median price.
Category 5: only occurs once, so no relevant information on 'typical' pricing; this advertiser will (normally) not occur in the test set either 
Catgories 1-4: occurs more than once, categorize based on median price 

df_test['advertiser'].value_counts(dropna = False).head(50) 

with open('intermediate_data/median_price_advertiser.pkl', 'rb') as file:
    median_price_advertiser = pickle.load(file)

df_test = pd.merge(df_test, median_price_advertiser, how='left', on='advertiser')
display(df_test)

df_test['advertiser_count'] = df_test['advertiser_count'].fillna(1)

df_test['advertiser_bin'].value_counts(dropna = False)

df_test['advertiser_bin'].value_counts(dropna = False)

df_test['advertiser_bin'] = np.where(pd.isna(df_test['advertiser_bin']), 5, df_test['advertiser_bin'])

df_test['advertiser_bin'].value_counts(dropna = False)

with open('intermediate_data/median_price_per_advertiser_bin.pkl', 'rb') as file:
    median_price_per_advertiser_bin = pickle.load(file)

df_test = pd.merge(df_test, median_price_per_advertiser_bin, how='left', on='advertiser_bin')
df_test.head()

## Regional prices - Statbel

with open('intermediate_data/map_nis_deduplicated.pkl', 'rb') as file:
    df_map_nis_deduplicated = pickle.load(file)

df_test = pd.merge(df_test, df_map_nis_deduplicated, how='left', left_on='postcode', right_on = 'zip_code')
display(df_test)

with open('intermediate_data/statbel_prices_selection.pkl', 'rb') as file:
    df_statbel_prices_selection = pickle.load(file)

df_test = pd.merge(df_test, df_statbel_prices_selection, how='left', left_on=['nis_code', 'is_appartment'], right_on = ['CD_REFNIS', 'F_APPARTMENT'])

display(df_test)

# Still quite some missings, whereas profiling below suggests it's a highly relevant feature

df_test['MS_P_50_median'].isna().sum()

with open('intermediate_data/price_lookup_zipcode.pkl', 'rb') as file:
    price_lookup_zipcode = pickle.load(file)

df_test["zip_code_first2"] = df_test.zip_code.astype(str).str[:2]

df_test["MS_P_50_median_imputed"] = df_test.apply(
    lambda row: price_lookup_zipcode.get(
        (row["zip_code_first2"], row["is_appartment"]),
        row["MS_P_50_median"]
    ) if pd.isna(row["MS_P_50_median"]) else row["MS_P_50_median"],
    axis=1
)

with open('intermediate_data/price_lookup_province.pkl', 'rb') as file:
    price_lookup_province = pickle.load(file)

df_test["zip_code_first1"] = df_test.zip_code.astype(str).str[:1]

df_test["MS_P_50_median_imputed"] = df_test.apply(
    lambda row: price_lookup_province.get(
        (row["zip_code_first1"], row["is_appartment"]),
        row["MS_P_50_median"]
    ) if pd.isna(row["MS_P_50_median_imputed"]) else row["MS_P_50_median_imputed"],
    axis=1
)

df_test[df_test['MS_P_50_median_imputed'].isna()]

## Regional prices - Price per area 

with open('intermediate_data/price_per_area_per_region.pkl', 'rb') as file:
    price_per_area_per_region = pickle.load(file)

df_test = pd.merge(df_test, price_per_area_per_region, how = 'left', on = 'zip_code_first2')

display(df_test)

## Lat / Lon

df_test['lat_missing'] = df_test['lat'].isna().astype(int)
df_test['lon_missing'] = df_test['lon'].isna().astype(int)

with open('intermediate_data/mean_lat.pkl', 'rb') as file:
    mean_lat = pickle.load(file)
with open('intermediate_data/mean_lon.pkl', 'rb') as file:
    mean_lon = pickle.load(file)

df_test['lat_imputed'] = df_test.apply(
    lambda row: mean_lat.get((row['province']), row['lat']) if pd.isna(row['lat']) else row['lat'],
    axis=1
)

df_test['lon_imputed'] = df_test.apply(
    lambda row: mean_lon.get((row['province']), row['lon']) if pd.isna(row['lon']) else row['lon'],
    axis=1
)

## Profile again

# Leave out percentiles 25 and 75 for prices per NIS code because of multicollinearity
# Leave out row count because it's not available when making a single prediction and has low predictive power (it's usually just 1)

# df_test_sel = df_test[['bedrooms', 'new_building', 'foto_amount', 'province', 'subtype_regrouped', 'area_missing', 'area_imputed', 'area_rel_to_bedrooms', 'energy_value_missing', 'energy_value_imputed', 'advertiser_count', 'median_price_advertiser_bin', 'MS_TOTAL_TRANSACTIONS', 'MS_P_50_median_imputed', 'price']].drop_duplicates()

df_test_sel = df_test[['id', 'new_building', 'foto_amount', 'province', 'subtype_median_price', 'area_missing', 'area_imputed', 'energy_value_missing', 'energy_value_imputed', 'advertiser_count', 'median_price_advertiser_bin', 'MS_P_50_median_imputed', 'median_price_per_area']].drop_duplicates()

profile_sel = ProfileReport(df_test_sel)

profile_sel.to_notebook_iframe()

## Final variable selection and encoding

display(df_test_sel)

# df_test_sel = df_test_sel[df_test_sel['price'] < 500000]

# X_test_preprocessed = df_test_sel.drop('price', axis = 1)
# X_test_preprocessed = pd.get_dummies(X_test_preprocessed, columns=['subtype_regrouped', 'province'], drop_first=True)

X_test_preprocessed = df_test_sel
X_test_preprocessed = pd.get_dummies(X_test_preprocessed, columns=['province'], drop_first=True)

display(X_test_preprocessed)

X_test_preprocessed.shape

X_test_preprocessed.to_csv("split_data/test_set_preprocessed.csv", index=False)

## Final variable selection and encoding - lat/lon instead of province

df_test_sel_2 = df_test[['id', 'new_building', 'foto_amount', 'lat_missing', 'lat_imputed', 'lon_missing', 'lon_imputed', 'subtype_median_price', 'area_missing', 'area_imputed', 'energy_value_missing', 'energy_value_imputed', 'advertiser_count', 'median_price_advertiser_bin', 'MS_P_50_median_imputed', 'median_price_per_area']].drop_duplicates()

X_test_preprocessed_2 = df_test_sel_2

X_test_preprocessed_2.to_csv("split_data/test_set_preprocessed_2.csv", index=False)
