==> 6_Predict_test_data.ipynb <==
import numpy as np
import pandas as pd
import pickle
from sklearn.ensemble import GradientBoostingRegressor
from mapie.regression import MapieQuantileRegressor

with open('models/gbr_sklearn_models.pkl', 'rb') as file:
    all_models = pickle.load(file)

X_test = pd.read_csv("split_data/test_set_preprocessed.csv")

display(X_test)

ID = X_test['id']
X_test = X_test.drop('id', axis = 1)

y_pred = all_models["q 0.50"].predict(X_test)
y_lower = all_models["q 0.05"].predict(X_test)
y_upper = all_models["q 0.95"].predict(X_test)
y_mid = (y_lower + y_upper)/2

# y_pred = all_models["mae"].predict(X_test) # does not always fall in interval

# ID,LOWER,UPPER,PRED

result = pd.DataFrame()
result["ID"] = ID
result["LOWER"] = y_lower.round(0).astype(int)
result["UPPER"] = y_upper.round(0).astype(int)
result["PRED"] = y_pred.round(0).astype(int)

display(result)

invalid_rows = result[~((result['PRED'] >= result['LOWER']) & (result['PRED'] <= result['UPPER']))]
display(invalid_rows)

result["MID"] = y_mid.round(0).astype(int)

result['PRED'] = np.where((result['PRED'] >= result['LOWER']) & (result['PRED'] <= result['UPPER']), result['PRED'], result['MID'])
result = result.drop('MID', axis = 1)
display(result)

invalid_rows = result[~((result['PRED'] >= result['LOWER']) & (result['PRED'] <= result['UPPER']))]
display(invalid_rows)

# Swap LOWER and UPPER where needed

result.loc[result['LOWER'] > result['UPPER'], ['LOWER', 'UPPER']] = result.loc[result['LOWER'] > result['UPPER'], ['UPPER', 'LOWER']].values

invalid_rows = result[~((result['PRED'] >= result['LOWER']) & (result['PRED'] <= result['UPPER']))]
display(invalid_rows)

display(result)

result.to_csv('result/aa1_42_attempt1.csv', index = False)

## With lat/lon instead of province

with open('models/gbr_sklearn_models_2.pkl', 'rb') as file:
    all_models = pickle.load(file)

X_test_2 = pd.read_csv("split_data/test_set_preprocessed_2.csv")

ID = X_test_2['id']
X_test_2 = X_test_2.drop('id', axis = 1)

y_pred = all_models["q 0.50"].predict(X_test_2)
y_lower = all_models["q 0.05"].predict(X_test_2)
y_upper = all_models["q 0.95"].predict(X_test_2)
y_mid = (y_lower + y_upper)/2

# ID,LOWER,UPPER,PRED

result = pd.DataFrame()
result["ID"] = ID
result["LOWER"] = y_lower.round(0).astype(int)
result["UPPER"] = y_upper.round(0).astype(int)
result["PRED"] = y_pred.round(0).astype(int)

invalid_rows = result[~((result['PRED'] >= result['LOWER']) & (result['PRED'] <= result['UPPER']))]
display(invalid_rows)

result["MID"] = y_mid.round(0).astype(int)

result['PRED'] = np.where((result['PRED'] >= result['LOWER']) & (result['PRED'] <= result['UPPER']), result['PRED'], result['MID'])
result = result.drop('MID', axis = 1)
display(result)

invalid_rows = result[~((result['PRED'] >= result['LOWER']) & (result['PRED'] <= result['UPPER']))]
display(invalid_rows)

# Swap LOWER and UPPER where needed

result.loc[result['LOWER'] > result['UPPER'], ['LOWER', 'UPPER']] = result.loc[result['LOWER'] > result['UPPER'], ['UPPER', 'LOWER']].values

invalid_rows = result[~((result['PRED'] >= result['LOWER']) & (result['PRED'] <= result['UPPER']))]
display(invalid_rows)

display(result)

result.to_csv('result/aa1_42_attempt2.csv', index = False)
